+++ fivetran/framework/resources/connector_schema_resource.go
@@ -69,6 +69,57 @@ func (r *connectorSchema) reloadSchema(ctx context.Context, schemaChangeHandling
 	return schemaResponse
 }
 
+func (r *connectorSchema) createNewSchema(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	client := r.GetClient()
+	if client == nil {
+		resp.Diagnostics.AddError(
+			"Unconfigured Fivetran Client",
+			"Please report this issue to the provider developers.",
+		)
+
+		return
+	}
+
+	var data model.ConnectorSchemaResourceModel
+
+	// Read Terraform plan data into the model
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
+
+	// Error while reading plan
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	// Plan is inconsistent
+	if !data.IsValid() {
+		resp.Diagnostics.AddError(
+			"Unable to Create Connector Schema Resource.",
+			"You can use solely one field to define schema settings.",
+		)
+		return
+	}
+
+	connectorID := data.ConnectorId.ValueString()
+	schemaChangeHandling := data.SchemaChangeHandling.ValueString()
+	localConfig := data.GetSchemaConfig()
+	svc := localConfig.PrepareCreateRequest(client.NewConnectorSchemaCreateService()).
+		ConnectorID(connectorID).
+		SchemaChangeHandling(schemaChangeHandling)
+
+	// we should not parse response here because it will contain only applied diffs, not the whole configuration
+	applyResponse, err := svc.Do(ctx)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Unable to Create Connector Schema Resource.",
+			fmt.Sprintf("Error while applying schema config patch. %v; code: %v; message: %v", err, applyResponse.Code, applyResponse.Message),
+		)
+		return
+	}
+	data.ReadFromResponse(applyResponse, &resp.Diagnostics)
+	data.Id = types.StringValue(connectorID)
+	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
+}
+
 func (r *connectorSchema) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 	if r.GetClient() == nil {
 		resp.Diagnostics.AddError(
@@ -84,15 +135,17 @@ func (r *connectorSchema) Create(ctx context.Context, req resource.CreateRequest
 	// Read Terraform plan data into the model
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
 
+	// Error while reading plan
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	// Plan is inconsistent
 	if !data.IsValid() {
 		resp.Diagnostics.AddError(
 			"Unable to Create Connector Schema Resource.",
 			"You can use solely one field to define schema settings.",
 		)
-		return
-	}
-
-	if resp.Diagnostics.HasError() {
 		return
 	}
 
@@ -108,12 +161,18 @@ func (r *connectorSchema) Create(ctx context.Context, req resource.CreateRequest
 		if schemaResponse.Code != "NotFound_SchemaConfig" {
 			resp.Diagnostics.AddError(
 				"Unable to Create Connector Schema Resource.",
-				fmt.Sprintf("Error wile retrieving existing schema. %v; code: %v; message: %v", err, schemaResponse.Code, schemaResponse.Message),
+				fmt.Sprintf("Error while retrieving existing schema. %v; code: %v; message: %v", err, schemaResponse.Code, schemaResponse.Message),
 			)
 			return
 		} else {
-			// reload because connector doens't have any schema settings yet
-			needReload = true
+			if data.ValidationLevel.ValueString() == "NONE" {
+				// create new desired schema
+				r.createNewSchema(ctx, req, resp)
+				return
+			} else {
+				// reload because connector doens't have any schema settings yet
+				needReload = true
+			}
 		}
 	} else {
 		// We might have to refresh schema, not all tables might be saved in current configuration
@@ -160,7 +219,7 @@ func (r *connectorSchema) Create(ctx context.Context, req resource.CreateRequest
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Unable to Create Connector Schema Resource.",
-			fmt.Sprintf("Error wile applying schema config patch. %v;", err),
+			fmt.Sprintf("Error while applying schema config patch. %v;", err),
 		)
 		return
 	}
@@ -179,7 +238,7 @@ func (r *connectorSchema) Create(ctx context.Context, req resource.CreateRequest
 		if err != nil {
 			resp.Diagnostics.AddError(
 				"Unable to Create Connector Schema Resource.",
-				fmt.Sprintf("Error wile applying schema config patch. %v; code: %v; message: %v", err, applyResponse.Code, applyResponse.Message),
+				fmt.Sprintf("Error while applying schema config patch. %v; code: %v; message: %v", err, applyResponse.Code, applyResponse.Message),
 			)
 			return
 		}
@@ -192,7 +251,7 @@ func (r *connectorSchema) Create(ctx context.Context, req resource.CreateRequest
 			if err != nil {
 				resp.Diagnostics.AddError(
 					"Unable to Create Connector Schema Resource.",
-					fmt.Sprintf("Error wile applying schema change handling policy. %v; code: %v; message: %v", err, schResponse.Code, schResponse.Message),
+					fmt.Sprintf("Error while applying schema change handling policy. %v; code: %v; message: %v", err, schResponse.Code, schResponse.Message),
 				)
 				return
 			}
@@ -221,7 +280,7 @@ func (r *connectorSchema) Create(ctx context.Context, req resource.CreateRequest
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Unable to Create Connector Schema Resource.",
-			fmt.Sprintf("Error wile applying schema config patch. %v.", err),
+			fmt.Sprintf("Error while applying schema config patch. %v.", err),
 		)
 		return
 	}
@@ -234,7 +293,7 @@ func (r *connectorSchema) Create(ctx context.Context, req resource.CreateRequest
 		if err != nil {
 			resp.Diagnostics.AddError(
 				"Unable to Create Connector Schema Resource.",
-				fmt.Sprintf("Error wile applying schema config patch. %v; code: %v; message: %v", err, applyResponse.Code, applyResponse.Message),
+				fmt.Sprintf("Error while applying schema config patch. %v; code: %v; message: %v", err, applyResponse.Code, applyResponse.Message),
 			)
 			return
 		}
@@ -286,7 +345,7 @@ func (r *connectorSchema) Read(ctx context.Context, req resource.ReadRequest, re
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Unable to Read Connector Schema Resource.",
-			fmt.Sprintf("Error wile retrieving existing schema. %v; code: %v; message: %v", err, schemaResponse.Code, schemaResponse.Message),
+			fmt.Sprintf("Error while retrieving existing schema. %v; code: %v; message: %v", err, schemaResponse.Code, schemaResponse.Message),
 		)
 		return
 	}
@@ -308,7 +367,13 @@ func (r *connectorSchema) Update(ctx context.Context, req resource.UpdateRequest
 	var plan, state model.ConnectorSchemaResourceModel
 
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
 	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
 
 	if !plan.IsValid() || !state.IsValid() {
 		resp.Diagnostics.AddError(
@@ -324,30 +389,27 @@ func (r *connectorSchema) Update(ctx context.Context, req resource.UpdateRequest
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Unable to Update Connector Schema Resource.",
-			fmt.Sprintf("Error wile retrieving existing schema. %v; code: %v; message: %v", err, schemaResponse.Code, schemaResponse.Message),
+			fmt.Sprintf("Error while retrieving existing schema settings. %v; code: %v; message: %v", err, schemaResponse.Code, schemaResponse.Message),
 		)
 		return
 	}
 
-	// Before applying changes we should validate existing state and planned changes and decide if we need to reload schema
-	planErr, _ := plan.ValidateSchemaElements(schemaResponse, *client, ctx)
-	stateErr, _ := state.ValidateSchemaElements(schemaResponse, *client, ctx)
-	if planErr != nil || stateErr != nil {
-		schemaResponse = r.reloadSchema(ctx, plan.SchemaChangeHandling.ValueString(), connectorID, resp.Diagnostics)
+	if plan.ValidationLevel.ValueString() != "NONE" {
+		// Before applying changes we should validate existing state and planned changes and decide if we need to reload schema
+		err, _ := plan.ValidateSchemaElements(schemaResponse, *client, ctx)
+		if err != nil {
+			schemaResponse = r.reloadSchema(ctx, plan.SchemaChangeHandling.ValueString(), connectorID, resp.Diagnostics)
+		}
+		err, _ = plan.ValidateSchemaElements(schemaResponse, *client, ctx)
+		if err != nil {
+			resp.Diagnostics.AddError(
+				"Unable to update Connector Schema Resource.",
+				fmt.Sprintf("Schema configuration is not aligned with source schema. Details:\n %v;", err),
+			)
+			return
+		}
 	}
 
-	err, _ = plan.ValidateSchemaElements(schemaResponse, *client, ctx)
-	if err != nil {
-		resp.Diagnostics.AddError(
-			"Unable to update Connector Schema Resource.",
-			fmt.Sprintf("Schema configuration is not aligned with source schema. Details:\n %v;", err),
-		)
-		return
-	}
-
-	if resp.Diagnostics.HasError() {
-		return
-	}
 	// read upstream config
 	config := configSchema.SchemaConfig{}
 	config.ReadFromResponse(schemaResponse)
@@ -361,7 +423,7 @@ func (r *connectorSchema) Update(ctx context.Context, req resource.UpdateRequest
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Unable to Create Connector Schema Resource.",
-			fmt.Sprintf("Error wile applying schema config patch. %v.", err),
+			fmt.Sprintf("Error while applying schema config patch. %v.", err),
 		)
 		return
 	}
@@ -380,7 +442,7 @@ func (r *connectorSchema) Update(ctx context.Context, req resource.UpdateRequest
 		if err != nil {
 			resp.Diagnostics.AddError(
 				"Unable to Create Connector Schema Resource.",
-				fmt.Sprintf("Error wile applying schema config patch. %v; code: %v; message: %v", err, applyResponse.Code, applyResponse.Message),
+				fmt.Sprintf("Error while applying schema config patch. %v; code: %v; message: %v", err, applyResponse.Code, applyResponse.Message),
 			)
 			return
 		}
@@ -394,7 +456,7 @@ func (r *connectorSchema) Update(ctx context.Context, req resource.UpdateRequest
 			if err != nil {
 				resp.Diagnostics.AddError(
 					"Unable to Update Connector Schema Resource.",
-					fmt.Sprintf("Error wile updating schema change handling policy. %v; code: %v; message: %v", err, schResponse.Code, schResponse.Message),
+					fmt.Sprintf("Error while updating schema change handling policy. %v; code: %v; message: %v", err, schResponse.Code, schResponse.Message),
 				)
 				return
 			}
@@ -421,7 +483,7 @@ func (r *connectorSchema) Update(ctx context.Context, req resource.UpdateRequest
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Unable to Update Connector Schema Resource.",
-			fmt.Sprintf("Error wile applying schema config patch. %v.", err),
+			fmt.Sprintf("Error while applying schema config patch. %v.", err),
 		)
 		return
 	}
@@ -433,7 +495,7 @@ func (r *connectorSchema) Update(ctx context.Context, req resource.UpdateRequest
 		if err != nil {
 			resp.Diagnostics.AddError(
 				"Unable to Create Connector Schema Resource.",
-				fmt.Sprintf("Error wile applying schema config patch. %v; code: %v; message: %v", err, applyResponse.Code, applyResponse.Message),
+				fmt.Sprintf("Error while applying schema config patch. %v; code: %v; message: %v", err, applyResponse.Code, applyResponse.Message),
 			)
 			return
 		}